various improvements + formatting
