+++ src/providers/gtfs-realtime.ts
@@ -132,23 +132,6 @@ export interface VehiclePositionOutput extends VehiclePosition {
     trip_headsign: string;
 }
 
-export function validateVehiclePosition(vehiclePosition: any): VehiclePositionOutput {
-    return {
-        rid: vehiclePosition.rid,
-        vid: vehiclePosition.vid,
-        lat: vehiclePosition.lat,
-        lon: vehiclePosition.lon,
-        heading: vehiclePosition.heading,
-        tripId: vehiclePosition.tripId,
-        stopIndex: vehiclePosition.stopIndex,
-        status: vehiclePosition.status,
-        trip_headsign: vehiclePosition.trip_headsign,
-        secsSinceReport: vehiclePosition.secsSinceReport,
-        stopId: vehiclePosition.stopId,
-        label: vehiclePosition.label,
-        delay: vehiclePosition.delay,
-    };
-}
 
 
 function makeVehicle(gtfsVehiclePosition, feedTimestamp, _vehicleIdKey): VehiclePosition {
